backend/contact/serializers.py
from rest_framework import serializers
from .models import Contact

class ContactSerializer(serializers.ModelSerializer):
    class Meta:
        model = Contact
        fields = ['id', 'name', 'email', 'service', 'message', 'created_at']
        read_only_fields = ['id', 'created_at'] 