backend/requirements.txt
Django==5.0.0
djangorestframework==3.14.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
sendgrid==6.10.0
django-cors-headers==4.3.1
gunicorn==21.2.0
python-decouple==3.8 